Added Account dao, domain model and mappers. Implemented account repository Fixed a type problem in passwordDao Implemented RoomAccountRepository Added sampleAccounts

+++ app/src/main/java/com/cr_d/passwordmanagerapp/data/daos/PasswordDao.kt
@@ -14,7 +14,7 @@ interface PasswordDao {
     suspend fun getAll(): List<PasswordEntity>
 
     @Query("SELECT * FROM PasswordEntity where id in (:userId)")
-    suspend fun getUserById(userId: Int): PasswordEntity?
+    suspend fun getUserById(userId: Long): PasswordEntity?
 
     @Insert(onConflict = OnConflictStrategy.ABORT)
     suspend fun insertPassword(password: PasswordEntity): Long
